Global changes. Removed ListOfChannelsActivity Added TicketsOverviewActivity Added SignUpActivity Added NavigationDrawer for them Added Sign Up button on homescreen Made first layouts for UserList and TicketList ListView Minor bugfixes
