+++ app/src/main/AndroidManifest.xml
@@ -27,13 +27,19 @@
         </activity>
 
         <activity
-            android:name=".ListOfChannelsActivity"
+            android:name=".CreateTicketActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
 
         <activity
-            android:name=".CreateTicketActivity"
+            android:name=".TicketsOverviewActivity"
+            android:label="@string/app_name"
+            android:theme="@style/AppTheme.NoActionBar">
+        </activity>
+
+        <activity
+            android:name=".SignUpUserActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
