+++ app/src/main/res/layout/activity_sign_up_user.xml
@@ -9,7 +9,7 @@
     tools:openDrawer="start">
 
     <include
-        layout="@layout/app_bar_channels"
+        layout="@layout/app_bar_sign_up_user"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
@@ -20,6 +20,6 @@
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
         app:headerLayout="@layout/nav_header_main"
-        app:menu="@menu/list_of_channels_drawer" />
+        app:menu="@menu/tickets_drawer" />
 
 </android.support.v4.widget.DrawerLayout>
