+++ app/src/main/res/layout/activity_tickets_overview.xml
@@ -9,7 +9,7 @@
     tools:openDrawer="start">
 
     <include
-        layout="@layout/app_bar_tickets_list"
+        layout="@layout/app_bar_tickets_overview"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
@@ -20,6 +20,6 @@
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
         app:headerLayout="@layout/nav_header_main"
-        app:menu="@menu/tickets_drawer" />
+        app:menu="@menu/tickets_overview_drawer" />
 
 </android.support.v4.widget.DrawerLayout>
