+++ app/src/main/res/layout/item_ticket.xml
@@ -1,6 +1,89 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1px"
+        android:background="#d9d9d9"
+        android:layout_alignParentBottom="true"
+        />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="horizontal"
+        android:paddingLeft="12dp"
+        android:paddingRight="12dp"
+        >
+        <ImageView
+            android:id="@+id/ticketImage"
+            android:layout_marginRight="17dp"
+            android:layout_width="50dp"
+            android:layout_height="50dp"
+            android:scaleType="centerCrop"
+            android:src="@mipmap/ic_launcher"
+            android:layout_gravity="center_vertical"
+            />
 
-</LinearLayout>
+        <RelativeLayout
+            android:layout_width="0dp"
+            android:layout_weight="1"
+            android:layout_height="wrap_content"
+            android:minHeight="50dp"
+            android:orientation="vertical"
+            android:layout_gravity="center_vertical"
+            >
+            <LinearLayout
+                android:id="@+id/top"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal"
+                >
+                <TextView
+                    android:id="@+id/ticketMember"
+                    android:layout_width="0dp"
+                    android:maxWidth="100dp"
+                    android:layout_weight="1"
+                    android:layout_height="wrap_content"
+                    android:singleLine="true"
+                    android:textSize="17sp"
+                    android:textColor="#3d3d3d"
+                    android:text="userName"
+                    />
+
+                <TextView
+                    android:id="@+id/ticketDate"
+                    android:layout_width="wrap_content"
+                    android:layout_height="match_parent"
+                    android:singleLine="true"
+                    android:textSize="10sp"
+                    android:textColor="#999999"
+                    android:text="Apr 16, 2015"
+                    android:layout_alignParentRight="true"
+                    android:gravity="top|center_horizontal"
+                    android:paddingTop="2dp"
+                    android:paddingLeft="10dp"
+                    />
+
+            </LinearLayout>
+            <TextView
+                android:layout_below="@+id/top"
+                android:id="@+id/ticketDesc"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:paddingTop="4dp"
+                android:maxLines="2"
+                android:textSize="12sp"
+                android:textColor="#999999"
+                android:text="Описание"
+                android:paddingRight="20dp" />
+        </RelativeLayout>
+    </LinearLayout>
+    <View
+        android:layout_alignParentBottom="true"
+        android:layout_width="match_parent"
+        android:layout_height="1px"
+        android:background="#e4e4e4"
+        />
+</RelativeLayout>
