+++ app/src/main/res/layout/item_user.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="76dp"
-    >
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
     <View
         android:layout_width="match_parent"
         android:layout_height="1px"
@@ -18,7 +17,7 @@
         android:paddingRight="12dp"
         >
         <ImageView
-            android:id="@+id/img_thumbnail"
+            android:id="@+id/userImage"
             android:layout_marginRight="17dp"
             android:layout_width="50dp"
             android:layout_height="50dp"
@@ -42,7 +41,7 @@
                 android:orientation="horizontal"
                 >
                 <TextView
-                    android:id="@+id/txt_topic"
+                    android:id="@+id/userName"
                     android:layout_width="0dp"
                     android:maxWidth="100dp"
                     android:layout_weight="1"
@@ -50,59 +49,21 @@
                     android:singleLine="true"
                     android:textSize="17sp"
                     android:textColor="#3d3d3d"
-                    android:text="Topic"
-                    />
-                <TextView
-                    android:id="@+id/txt_member_count"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:background="@drawable/icon_group_number"
-                    android:text="1"
-                    android:textSize="10sp"
-                    android:textColor="#a4acbc"
-                    android:layout_gravity="center_vertical"
-                    android:gravity="center"
-                    android:layout_marginLeft="6dp"
-                    />
-                <TextView
-                    android:id="@+id/txt_date"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:singleLine="true"
-                    android:textSize="10sp"
-                    android:textColor="#999999"
-                    android:text="Apr 16, 2015"
-                    android:layout_alignParentRight="true"
-                    android:gravity="top|center_horizontal"
-                    android:paddingTop="2dp"
-                    android:paddingLeft="10dp"
+                    android:text="userName"
                     />
 
             </LinearLayout>
             <TextView
                 android:layout_below="@+id/top"
-                android:id="@+id/txt_desc"
+                android:id="@+id/userDate"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:paddingTop="4dp"
                 android:maxLines="2"
                 android:textSize="12sp"
                 android:textColor="#999999"
-                android:text="Последнее сообщение."
-                android:paddingRight="20dp"
-                />
-            <TextView
-                android:layout_below="@+id/top"
-                android:id="@+id/txt_unread_count"
-                android:layout_width="25dp"
-                android:layout_height="25dp"
-                android:layout_alignParentRight="true"
-                android:text="99"
-                android:textColor="#ffffff"
-                android:textSize="12sp"
-                android:gravity="center"
-                android:background="@drawable/bg_notify"
-                />
+                android:text="Дата подачи заявки"
+                android:paddingRight="20dp" />
         </RelativeLayout>
     </LinearLayout>
     <View
