+++ app/src/main/res/menu/tickets_drawer.xml
@@ -5,7 +5,11 @@
         <item
             android:id="@+id/listOfChannels"
             android:icon="@drawable/ic_message_black_24dp"
-            android:title="Список каналов" />
+            android:title="Список текущих заявок" />
+        <item
+            android:id="@+id/signUpUser"
+            android:icon="@drawable/ic_assignment_turned_in_black_24dp"
+            android:title="Авторизация пользователей" />
         <item
             android:id="@+id/settings"
             android:icon="@drawable/ic_settings_black_24dp"
