+++ app/src/main/res/menu/tickets_overview_drawer.xml
@@ -5,7 +5,11 @@
         <item
             android:id="@+id/listOfTickets"
             android:icon="@drawable/ic_report_black_24dp"
-            android:title="Заявки" />
+            android:title="Список возможных заявок" />
+        <item
+            android:id="@+id/signUpUser"
+            android:icon="@drawable/ic_assignment_turned_in_black_24dp"
+            android:title="Авторизация пользователей" />
         <item
             android:id="@+id/settings"
             android:icon="@drawable/ic_settings_black_24dp"
